Move some URL related helpers to the new helper class
And deprecate the old methods.

Fixes: OX-11739
